CH3/coupon-core/src/main/java/com/yb/couponcore/service/CouponIssueRedisService.java
package com.yb.couponcore.service;

import com.yb.couponcore.repository.redis.RedisRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import static com.yb.couponcore.util.CouponRedisUtils.getIssueRequestKey;

@Service
@RequiredArgsConstructor
public class CouponIssueRedisService {

    private final RedisRepository redisRepository;

    public boolean availableTotalIssueQuantity(Integer totalQuantity , long couponId){
        if(totalQuantity == null){
            return true;
        }

        String key = getIssueRequestKey(couponId);
        return totalQuantity > redisRepository.sCard(key);
    }

    public boolean availableUserIssueQuantity(long couponId , long userId){
        String key = getIssueRequestKey(couponId);
        return !redisRepository.sIsMember(key , String.valueOf(userId));
    }

}
